cleanup and move user input logix into servoshell
