+++ ports/servoshell/Cargo.toml
@@ -56,7 +56,7 @@ glow = "0.12.2"
 keyboard-types = { workspace = true }
 lazy_static = { workspace = true }
 libc = { workspace = true }
-libservo = { path = "../../components/servo" }
+libservo = { path = "../../components/servo", features=["media-gstreamer"] }
 log = { workspace = true }
 raw-window-handle = "0.5"
 servo-media = { git = "https://github.com/servo/media" }
@@ -65,6 +65,7 @@ surfman = { workspace = true, features = ["sm-x11", "sm-raw-window-handle"] }
 tinyfiledialogs = "3.0"
 webxr = { git = "https://github.com/servo/webxr", features = ["ipc", "glwindow", "headless"] }
 winit = "0.28.6"
+url = { workspace = true }
 
 [target.'cfg(any(target_os = "linux", target_os = "windows"))'.dependencies]
 image = { workspace = true }
